+++ lib/src/route_processor.dart
@@ -15,6 +15,7 @@ import 'parameters_processor.dart';
 import 'metadata.dart';
 import 'response_writer.dart';
 import 'dynamic_map.dart';
+import 'constants.dart';
 
 class RouteProcessor implements Function {